Authentication finished, routes protected

+++ exercise_tracker_prototype/frontend/src/components/WorkoutDetails.js
@@ -1,4 +1,5 @@
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
+import { useAuthContext } from "../hooks/useAuthContext";
 import formatDistanceToNow from 'date-fns/formatDistanceToNow';
 
 // function formatDistnaceToNow(date, options){
@@ -7,10 +8,17 @@ import formatDistanceToNow from 'date-fns/formatDistanceToNow';
 
 const WorkoutDetails = ( {workout} ) => {
     const { dispatch } = useWorkoutsContext();
+    const { user } = useAuthContext();
 
     const handleClick = async (req, res) => {
+        if(!user){
+            return;
+        }
         const response = await fetch('/server/api/workouts/' + workout._id, {
-            method: "DELETE"
+            method: "DELETE",
+            headers: {
+                "Authorization": `Bearer ${user.token}` 
+            }
         });
         const json = await response.json();
